Configure bindgen for cross compiling
This is a lot of noise for not much difference - it just means that
tflite_types is written for 32-bit pointers and alignments rather than
64-bit when cross compiling.

Unfortunately libclang can't figure out where the cross c++ standard
libraries are on its own. So we have to run our cross compiler with -v,
scrape the output (presuming its gcc) and then pass this to
libclang. Scraping is todo, just hardcoded for arm-none-eabi-gcc for now.